Add pallet.test.image-list to select images to use for testing
When running the live tests, it is useful to be able to specify which images
to use.  The pallet.test.image-list can be used to specify this. The list
of images is made available to the tests in pallet.live-test/*images*.  The
pallet.live-test/exclude-images and pallet.live-test/filter-images can be
used to special case tests, or filter unsupported images.