Fix bug in pallet.script/add-to-scripts
The add-to-scripts function was referencing *scripts*, which was incorrect
since it is called by an alter-var-root invoctation on *scripts*.

Added extra tests and logging around script template selection for
generation of the bootstrap script.

Added a condition/raise when jclouds returns OsFamily/UNRECOGNISED on the
image to be used for a new node.