Updated implementation of for->, with support for destructuring.
The existing implementation of for-> supported the binding of a single
variable to a single sequence. The updated implementation here makes
use of clojure.core/for to build up a sequence of functions;
((apply comp functionseq) arg) is then used to thread the argument
among the sequences. Each item in the resulting sequence is dealt with
in order.