Modification to phase-fn to allow for an argument vector.
Implemented phase-fn using recursion to allow for very specific
debugging messages -- the threaded session is now able to report the
subphase into which it was going to be threaded when check-session
returns an error. Errors will almost certainly be found in the
previous expression.

Changed all current tests to support a zero-argument vector.