Added macro-writing macro that allows for custom definition of functions similar to `phase-fn`, with the added benefit that `defthreadfn` allows for any number of checkers to be placed after each threaded argument.
For example,

    (defthreadfn ultrasafe-phase-fn
        "docstring!"
         watchdog-func-1 watchdog-func-2 watchdog-func-3)

Will define a macro that can be called like this:

    (ultrasafe-phase-fn [x y z]
        (phase1 x y)
        (phase2 x y z))

Which will (recursively) expand into

    (fn [arg x y z]
      (--> arg
           (phase1 x y)
           (watchdog-func-1 "(phase1 x y)")
           (watchdog-func-2 "(phase1 x y)")
           (watchdog-func-3 "(phase1 x y)")
           (phase2 x y z)
           (watchdog-func-1 "(phase2 x y z)")
           (watchdog-func-2 "(phase2 x y z)")
           (watchdog-func-3 "(phase2 x y z)")))

`-->`, from a prior commit in the pull request, allows for the use of
`for`, `when`, let` and `expose-request-as` within the body of the
function. What's nice about this construction is that

    (defthreadfn unchecked-phase-fn)

When called, expands to

    (fn [arg x y z]
      (--> arg
           (phase1 x y)
           (phase2 x y z)

Which simply gives access to `-->`, with no checking.

Currently, defthreadfn accepts attr-maps and docstrings, but doesn't
format its attr-map well. I'll look at this for the next commit --
this macro is debugged, so I figured I'd submit it for your initial
comments.