Make image selection from template more deterministic.
The metadata of the image is now stored in the VM as an extended parameter.
Image selection is now based only on predefined selectors matching the ones in jclouds and will ignore any other entry.
find-images will find all the images that match a template.
The image metadata is now preprocessed to add some fields automatically for compatibility with jclouds.