Add on-one-node macro
The body of the on-one-macro will be executed on only one node out of
all the nodes that have the given roles. If more than one role is
specified, and there is at least one node that provides the intersection
of all the roles, then first of any such nodes is used.