Adding support for Travis builds
This change includes a .travis.yml file and travis/settings.xml file.  The settings.xml file is used by travis to build pallet since pallet relies upon your ~/settings.xml file to be set-up a certain way (i.e. not all of it is in the pom).  The .travis file contains the code to build it using travis.  The basics are it builds and then runs the tests.  Finally, the README.md file was updated to denote status of the current build for this branch