Refactor etc-hosts crate
The host entries are now maintained in the settings, with functions to add
localhost, ipv6-aliases and a named localhost-hostname (using 127.0.1.1). The
aliases for each ip are now stored in a sequence.

The new add-hosts function allows for bulk update of the hosts mapping.

The hosts-for-group and hosts-for role are removed.  Equivalent functionality
can be achieved using pallet.crate/nodes-with-role and the add-hosts function.