Add pallet.repl functions for easier crate development
Add show-nodes, show-group for pretty printing the nodes in a provider
or in a group within a provider

Add explain-plan to show (print) what pallet is going to do with a
plan function, both in terms of generated actions forms and also shell
scripts.

Add pallet.core.data-api to hold the data-only version of such
functions.