Support node files for the node-list provider
Support instantiating a node-list provider from clojure format data files.

Introduces pallet.compute.node-list/node, which is similar to make-node, but
makes the os-family, name and group name optional.  The name and group-name
default to the ip.  This means the only data required is an ip address.

Adds a :node-file option to the provide function.  If neither :node-list nor
:node-file are specified, it looks for a node file in standard locations:
.pallet-nodes, ~/.pallet/nodes and /etc/pallet/nodes.

Introduces pallet.compute.node-list/node-list, which is similar to
node-list-service, but makes the node-list argument optional.