Remove two phase execution (and the action-plan)
This removes the two phase execution, which removes the need for an
action plan.  The action plan was used for aggregation of actions,
such as package, which is better achieved via the settings phase.

The automatic removal of duplicate actions now no longer takes place.

There are no precedence relations between actions.